requirements.txt
requests
jinja2
google-generativeai
python-dotenv
